blip_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
numpy==1.26.4
opencv-python-headless==4.10.0.84
pillow==10.4.0
torch==2.2.2
torchvision==0.17.2
ultralytics==8.3.0
transformers==4.57.3
safetensors==0.4.5
accelerate==0.34.0
